requirements.txt
fastapi
uvicorn
requests
python-dotenv
httpx
pandas
oauth2client